fix: Resolve CI lint and static analysis errors
Fixes code style issue in StatsServiceCacheTest.php (new_with_braces).
Fixes return type mismatch in FetchBodyPartMeasurementsIndexAction.php by explicitly casting round() result to float.
Verified with pint, phpstan, and feature tests.

+++ app/Actions/Measurements/FetchBodyPartMeasurementsIndexAction.php
@@ -60,7 +60,7 @@ final class FetchBodyPartMeasurementsIndexAction
                 'current' => $latest->value,
                 'unit' => $latest->unit,
                 'date' => Carbon::parse($latest->measured_at)->format('Y-m-d'),
-                'diff' => $previous ? round($latest->value - $previous->value, 2) : 0,
+                'diff' => $previous ? (float) round($latest->value - $previous->value, 2) : 0.0,
             ];
         })->values();
     }
